+++ test/playwright/radios.spec.ts
@@ -1,8 +1,29 @@
 import { test, expect } from "@playwright/test";
-import { expectFormFailure, expectFormSuccess } from "./lib";
+import {
+  expectFormFailure,
+  expectFormSuccess,
+  expectSingleFieldValue,
+} from "./lib";
 
 test("radios", async ({ page }) => {
   await page.goto("/forms/radios");
+  await expect(await page.getByTestId("form")).toMatchAriaSnapshot(`
+  - group "Level":
+    - text: Level
+    - radio "Apprentice"
+    - text: Apprentice
+    - radio "Junior"
+    - text: Junior
+    - radio "Mid-level"
+    - text: Mid-level
+    - radio "Senior"
+    - text: Senior
+    - radio "Lead"
+    - text: Lead
+    - radio "Principal"
+    - text: Principal
+  - button "Submit"`);
+  await expectSingleFieldValue(page, null);
   await page.getByRole("button", { name: "Submit" }).click();
 
   await expectFormFailure(page);
@@ -13,6 +34,7 @@ test("radios", async ({ page }) => {
   await expect(page.getByLabel("Senior")).not.toBeChecked();
   await expect(page.getByLabel("Lead")).not.toBeChecked();
   await expect(page.getByLabel("Principal")).not.toBeChecked();
+  await expectSingleFieldValue(page, null);
   await page.getByLabel("Senior").check();
   await page.getByRole("button", { name: "Submit" }).click();
 
@@ -23,4 +45,5 @@ test("radios", async ({ page }) => {
   await expect(page.getByLabel("Senior")).toBeChecked();
   await expect(page.getByLabel("Lead")).not.toBeChecked();
   await expect(page.getByLabel("Principal")).not.toBeChecked();
+  await expectSingleFieldValue(page, "4");
 });